refactor(artquiz): removes context
