+++ src/Apps/ArtQuiz/Routes/ArtQuizArtworks.tsx
@@ -1,4 +1,3 @@
-import { ArtQuizContextProvider } from "Apps/ArtQuiz/ArtQuizContext"
 import { ArtQuizMain } from "Apps/ArtQuiz/Components/ArtQuizMain"
 import { MetaTags } from "Components/MetaTags"
 import { FC } from "react"
@@ -8,9 +7,7 @@ export const ArtQuizArtworks: FC = () => {
     <>
       <MetaTags title="Art Taste Quiz | Artsy" />
 
-      <ArtQuizContextProvider>
-        <ArtQuizMain />
-      </ArtQuizContextProvider>
+      <ArtQuizMain />
     </>
   )
 }
